requirements.txt
flask==3.0.0
google-generativeai==0.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
python-multipart==0.0.6
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
aiohttp==3.9.1
asyncio==3.4.3
uvloop==0.19.0.1
